reorganize existing crypto unit tests into their own files
